Iterator Perfomance Improvements (#3667)
* steal benches


* Gen floats, ints and bool predicates


* remove string allocs


* added gen to make


* Gen strings


* patch up vp2 and vp4


* remove instrumented predicate


* int32 rns


* skip nulls


* vp4 unsafe string


* revert isnull check due to impact on query range


* remove string range cond


* changelog


* lint and tests


* exclude


* again


* fix second line


* fix bool pred


* change generated file name and exclude


* additional tests


---------

+++ tempodb/encoding/vparquet4/block_findtracebyid.go
@@ -246,7 +246,7 @@ func findTraceByID(ctx context.Context, traceID common.ID, maxTraceSizeBytes int
 	for _, rg := range pf.RowGroups()[0:rowGroup] {
 		rowMatch += rg.NumRows()
 	}
-	rowMatch += res.RowNumber[0]
+	rowMatch += int64(res.RowNumber[0])
 
 	// seek to row and read
 	r := parquet.NewGenericReader[*Trace](pf)
